refactor: organize Socket.IO constants into a dedicated module and update imports
