+++ services/web/server/src/simcore_service_webserver/socketio/socketio_service.py
@@ -3,12 +3,6 @@ from ._utils import (
     register_socketio_handlers,
 )
 from .messages import (
-    SOCKET_IO_EVENT,
-    SOCKET_IO_HEARTBEAT_EVENT,
-    SOCKET_IO_LOG_EVENT,
-    SOCKET_IO_NODE_UPDATED_EVENT,
-    SOCKET_IO_PROJECT_UPDATED_EVENT,
-    SOCKET_IO_WALLET_OSPARC_CREDITS_UPDATED_EVENT,
     send_message_to_project_room,
     send_message_to_standard_group,
     send_message_to_user,
@@ -16,13 +10,6 @@ from .messages import (
 from .plugin import setup_socketio
 
 __all__: tuple[str, ...] = (
-    # constants
-    "SOCKET_IO_EVENT",
-    "SOCKET_IO_HEARTBEAT_EVENT",
-    "SOCKET_IO_LOG_EVENT",
-    "SOCKET_IO_NODE_UPDATED_EVENT",
-    "SOCKET_IO_PROJECT_UPDATED_EVENT",
-    "SOCKET_IO_WALLET_OSPARC_CREDITS_UPDATED_EVENT",
     # functions
     "get_socket_server",
     "register_socketio_handlers",
