+++ services/web/server/src/simcore_service_webserver/users/models.py
@@ -1,6 +1,6 @@
-# mypy: disable-error-code=truthy-function
-
 from ._models import FullNameDict, UserDisplayAndIdNamesTuple
 
-__all__: tuple[str, ...] = ("FullNameDict", "UserDisplayAndIdNamesTuple")
-# nopycln: file
+__all__: tuple[str, ...] = (
+    "FullNameDict",
+    "UserDisplayAndIdNamesTuple",
+)  # nopycln: file
